// src/app/@core/model/transaksi.model.ts
export class TransaksiModel{
    id: number;
    id_transaksi: string;
    metode: string;
    kontak:string
    nama_tempat_wisata: string;
    tanggal: any;
    transaksi: any;
    no_transaksi: string;
    no_telp: string;
    email: string;
    total_bayar: any;
    status_bayar: number;
    metode_pembayaran: string;
    id_tempat_wisata: number;
}